+++ src/lib.rs
@@ -5,14 +5,14 @@
 //!
 //! ## Motivation
 //!
-//! `presser` can help you when copying data into raw buffers. One primary use-case is copying data into
+//! `presser` can help you when copying data into and reading data out of raw buffers. One primary use-case is copying data into
 //! graphics-api-allocated buffers which will then be accessed by the GPU. Common methods for doing this
 //! right now in Rust can often invoke UB in subtle and hard-to-see ways. For example, viewing an allocated
 //! but uninitialized buffer as an `&mut [u8]` **is instantly undefined behavior**\*, and `transmute`ing even a
 //! `T: Copy` type which has *any padding bytes in its layout* as a `&[u8]` to be the source of a copy is
 //! **also instantly undefined behavior**, in both cases because it is *invalid* to create a reference to an invalid
 //! value (and uninitialized memory is an invalid `u8`), *even if* your code never actually accesses that memory.
-//! This immediately makes what seems like the most straightforward way to copy data into buffers unsound 😬.
+//! This immediately makes what seems like the most straightforward way to copy data into buffers unsound 😬
 //!
 //! `presser` helps with this by allowing you to view raw allocated memory of some size as a "[`Slab`]" of memory and then
 //! provides *safe, valid* ways to copy data into that memory. For example, you could implement [`Slab`] for your
@@ -24,10 +24,10 @@
 //! \* *If you're currently thinking to yourself "bah! what's the issue? surely an uninit u8 is just any random bit pattern
 //! and that's fine we don't care," [check out this blog post](https://www.ralfj.de/blog/2019/07/14/uninit.html) by
 //! @RalfJung, one of the people leading the effort to better define Rust's memory and execution model. As is explored
-//! in that blog post, an *uninit* piece of memory is not simply *an arbitrary bit pattern*, it is a wholly separate
+//! in that blog post, an* uninit *piece of memory is not simply* an arbitrary bit pattern, *it is a wholly separate
 //! state about a piece of memory, outside of its value, which lets the compiler perform optimizations that reorder,
 //! delete, and otherwise change the actual execution flow of your program in ways that cannot be described simply
-//! by "the value could have *some* possible bit pattern". LLVM and Clang are changing themselves to require special
+//! by "the value could have* some *possible bit pattern". LLVM and Clang are changing themselves to require special
 //! `noundef` attribute to perform many important optimizations that are otherwise unsound. For a concrete example
 //! of the sorts of problems this can cause,
 //! [see this issue @scottmcm hit](https://github.com/rust-lang/rust/pull/98919#issuecomment-1186106387).*
@@ -44,7 +44,7 @@
 //! a [`BorrowedRawAllocation`] (which implements [`Slab`]) by calling the unsafe function
 //! [`RawAllocation::borrow_as_slab`]
 //!
-//! Once you have a slab, you can use the copy helper functions provided at the crate root, for example,
+//! Once you have a slab, you can use the helper functions provided at the crate root, for example,
 //! [`copy_to_offset`] and [`copy_to_offset_with_align`].
 //!
 //! ### Example
@@ -78,6 +78,13 @@
 //! // `my_data` may be placed at a different offset than requested. so, we check the returned
 //! // `CopyRecord` to check the actual start offset of the copied data.
 //! let actual_start_offset = copy_record.copy_start_offset;
+//!
+//! // we may later (*unsafely*) read back our data. note that the read helpers provided by presser
+//! // are mostly unsafe. They do help protect you from some common footguns, but you still ultimately need
+//! // to guarantee you put the proper data where you're telling it you put the proper data.
+//! let my_copied_data_in_my_buffer: &MyDataStruct = unsafe {
+//!     presser::read_at_offset(&slab, actual_start_offset)?
+//! };
 //! ```
 //!
 //! ### `#[no_std]`
@@ -94,13 +101,16 @@
 //! using same backing memory that the [`Slab`] you copied into used are still safe.
 //!
 //! For example, say you have a fully-initialized
-//! chunk of bytes (like a `Vec<u8>`), which you view as a [`Slab`], and then (safely) perform a copy
+//! chunk of bytes (like a `Vec<u8>`), which you (unsafely\*) view as a [`Slab`], and then (safely) perform a copy
 //! operation into using [`copy_to_offset`]. If the `T` you copied into it has any padding bytes in
 //! its memory layout, then the memory locations where those padding bytes now exist in the underlying `Vec`'s
 //! memory must now be treated as uninitialized. As such, taking any view into that byte vector which
 //! relies on those newly-uninit bytes being initialized to be valid (for example, taking a `&[u8]` slice of the `Vec`
 //! which includes those bytes, ***even if your code never actually reads from that slice***)
 //! is now instant **undefined behavior**.
+//!
+//! \* *Note: this is unsafe because, as exemplified, you may copy uninit data into the buffer. Hence, care should
+//! be taken when implementing [`Slab`] and then providing a safe interface on top of a low level buffer type.*
 #![cfg_attr(not(feature = "std"), no_std)]
 #![deny(unsafe_op_in_unsafe_fn)]
 #![deny(missing_docs)]
@@ -117,8 +127,14 @@ use core::marker::PhantomData;
 use core::mem::MaybeUninit;
 use core::ptr::NonNull;
 
+mod copy;
+mod read;
+
+pub use copy::*;
+pub use read::*;
+
 /// Represents a contiguous piece of a single allocation with some layout that is used as a
-/// data copying destination. May be wholly or partially uninitialized.
+/// data copying destination or reading source. May be wholly or partially uninitialized.
 ///
 /// This trait is *basically* equivalent to implementing `Deref`/`DerefMut` with
 /// `Target = [MaybeUninit<u8>]` in terms of safety requirements. It is a separate
@@ -288,6 +304,43 @@ unsafe impl Slab for [MaybeUninit<u8>] {
 
     fn size(&self) -> usize {
         core::mem::size_of_val(self)
+    }
+}
+
+/// An error that may occur during a copy or read operation.
+#[derive(Debug)]
+pub enum Error {
+    /// Copy or read would exceed the end of the allocation
+    OutOfMemory,
+    /// Requested to copy to or read from an offset outside the bounds of the allocation
+    OffsetOutOfBounds,
+    /// Computed invalid layout for copy operation, probably caused by incredibly large size, offset, or min-alignment parameters
+    InvalidLayout,
+    /// The requested offset was unalignd. In a read operation, this means the provided offset into the buffer was not properly aligned
+    /// for the requested type.
+    ///
+    /// In an `exact` variant copy function, the computed copy start offset did not match the requested start offset,
+    /// meaning the requested start offset was not properly aligned.
+    RequestedOffsetUnaligned,
+}
+
+impl core::fmt::Display for Error {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(f, "{}", match self {
+            Self::OutOfMemory => "End of copy or read operation would exceed the end of the allocation",
+            Self::OffsetOutOfBounds => "Requested read from or copy to a location starting outside the allocation",
+            Self::InvalidLayout => "Computed invalid layout requirements, probably caused by incredibly large size, offset, or alignment parameters",
+            Self::RequestedOffsetUnaligned => "Requested offset into Slab did not satisfy computed alignment requirements",
+        })
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for Error {}
+
+impl From<LayoutError> for Error {
+    fn from(_err: LayoutError) -> Self {
+        Self::InvalidLayout
     }
 }
 
@@ -477,6 +530,56 @@ unsafe impl<'a> Slab for BorrowedRawAllocation<'a> {
     }
 }
 
+/// Computed offsets necessary for a copy or read operation with some layout. Should only be
+/// created by [`compute_offsets`]
+#[derive(Debug, Copy, Clone)]
+pub(crate) struct ComputedOffsets {
+    start: usize,
+    end: usize,
+    end_padded: usize,
+}
+
+/// Compute and validate offsets for a copy or read operation with the given parameters.
+pub(crate) fn compute_and_validate_offsets<S: Slab>(
+    slab: &S,
+    start_offset: usize,
+    t_layout: Layout,
+    min_alignment: usize,
+    require_exact_start_offset: bool,
+) -> Result<ComputedOffsets, Error> {
+    let layout = t_layout.align_to(min_alignment.next_power_of_two())?;
+
+    let computed_start_offset =
+        align_offset_up_to(slab.base_ptr() as usize, start_offset, layout.align())
+            .ok_or(Error::InvalidLayout)?;
+    if require_exact_start_offset && start_offset != computed_start_offset {
+        return Err(Error::RequestedOffsetUnaligned);
+    }
+    let computed_end_offset = computed_start_offset
+        .checked_add(layout.size())
+        .ok_or(Error::InvalidLayout)?;
+    let computed_end_offset_padded = computed_start_offset
+        .checked_add(layout.pad_to_align().size())
+        .ok_or(Error::InvalidLayout)?;
+
+    // check start is inside slab
+    // if within slab, we also know that copy_start_offset is <= isize::MAX since slab.size() must be <= isize::MAX
+    if computed_start_offset > slab.size() {
+        return Err(Error::OffsetOutOfBounds);
+    }
+
+    // check end is inside slab
+    if computed_end_offset_padded > slab.size() {
+        return Err(Error::OutOfMemory);
+    }
+
+    Ok(ComputedOffsets {
+        start: computed_start_offset,
+        end: computed_end_offset,
+        end_padded: computed_end_offset_padded,
+    })
+}
+
 /// Given pointer and offset, returns a new offset aligned to `align`.
 ///
 /// `align` *must* be a power of two and >= 1 or else the result is meaningless.
@@ -485,488 +588,4 @@ fn align_offset_up_to(ptr: usize, offset: usize, align: usize) -> Option<usize> 
     let aligned_ptr = offsetted_ptr.checked_add(align - 1)? & !(align - 1);
     // don't need to check since we know aligned_ptr is >= ptr at this point
     Some(aligned_ptr - ptr)
-}
-
-/// Compute and validate offsets for a copy operation with the given parameters.
-fn compute_offsets<S: Slab>(
-    dst: &S,
-    start_offset: usize,
-    t_layout: Layout,
-    min_alignment: usize,
-    require_exact_start_offset: bool,
-) -> Result<CopyRecord, CopyError> {
-    let copy_layout = t_layout.align_to(min_alignment.next_power_of_two())?;
-
-    let copy_start_offset =
-        align_offset_up_to(dst.base_ptr() as usize, start_offset, copy_layout.align())
-            .ok_or(CopyError::InvalidLayout)?;
-    if require_exact_start_offset && start_offset != copy_start_offset {
-        return Err(CopyError::RequestedOffsetUnaligned);
-    }
-    let copy_end_offset = copy_start_offset
-        .checked_add(copy_layout.size())
-        .ok_or(CopyError::InvalidLayout)?;
-    let copy_end_offset_padded = copy_start_offset
-        .checked_add(copy_layout.pad_to_align().size())
-        .ok_or(CopyError::InvalidLayout)?;
-
-    // check start is inside slab
-    // if within slab, we also know that copy_start_offset is <= isize::MAX since slab.size() must be <= isize::MAX
-    if copy_start_offset > dst.size() {
-        return Err(CopyError::OffsetOutOfBounds);
-    }
-
-    // check end is inside slab
-    if copy_end_offset_padded > dst.size() {
-        return Err(CopyError::OutOfMemory);
-    }
-
-    Ok(CopyRecord {
-        copy_start_offset,
-        copy_end_offset,
-        copy_end_offset_padded,
-    })
-}
-
-/// An error that may occur during a copy operation.
-#[derive(Debug)]
-pub enum CopyError {
-    /// Copy would exceed the end of the allocation
-    OutOfMemory,
-    /// Requested to copy to an offset outside the bounds of the allocation
-    OffsetOutOfBounds,
-    /// Computed invalid layout for copy operation, probably caused by incredibly large size, offset, or min-alignment parameters
-    InvalidLayout,
-    /// In an `exact` variant copy function, the computed copy start offset did not match the requested start offset,
-    /// meaning the requested start offset was not properly aligned
-    RequestedOffsetUnaligned,
-}
-
-impl core::fmt::Display for CopyError {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{}", match self {
-            Self::OutOfMemory => "Copy would exceed the end of the allocation",
-            Self::OffsetOutOfBounds => "Requested copy to a location starting outside the allocation",
-            Self::InvalidLayout => "Invalid layout, probably caused by incredibly large size, offset, or alignment parameters",
-            Self::RequestedOffsetUnaligned => "Requested start_offset did not satisfy computed alignment requirements",
-        })
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for CopyError {}
-
-impl From<LayoutError> for CopyError {
-    fn from(_err: LayoutError) -> Self {
-        Self::InvalidLayout
-    }
-}
-
-/// Record of the results of a copy operation
-#[derive(Debug, Copy, Clone)]
-pub struct CopyRecord {
-    /// The offset from the start of the allocation, in bytes, at which the
-    /// copy operation began to write data.
-    ///
-    /// Not necessarily equal to the `start_offset`, since this offset
-    /// includes necessary padding to assure alignment.
-    pub copy_start_offset: usize,
-
-    /// The offset from the start of the allocation, in bytes, at which the
-    /// copy operation no longer wrote data.
-    ///
-    /// This does not include any padding at the end necessary to maintain
-    /// alignment requirements.
-    pub copy_end_offset: usize,
-
-    /// The offset from the start of the allocation, in bytes, at which the
-    /// copy operation no longer wrote data, plus any padding necessary to
-    /// maintain derived alignment requirements.
-    pub copy_end_offset_padded: usize,
-}
-
-/// Copies `src` into the memory represented by `dst` starting at *exactly*
-/// `start_offset` bytes past the start of `dst`
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`, in bytes,
-/// where the first byte of the copied data will be placed. If the requested
-/// start offset does not satisfy computed alignment requirements, an error will
-/// be returned and no data will be copied.
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_to_offset_exact<T: Copy, S: Slab>(
-    src: &T,
-    dst: &mut S,
-    start_offset: usize,
-) -> Result<CopyRecord, CopyError> {
-    copy_to_offset_with_align_exact(src, dst, start_offset, 1)
-}
-
-/// Copies `src` into the memory represented by `dst` starting at *exactly*
-/// `start_offset` bytes past the start of `dst` and with minimum alignment
-/// `min_alignment`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`, in bytes,
-/// where the first byte of the copied data will be placed. If the requested
-/// start offset does not satisfy computed alignment requirements, an error will
-/// be returned and no data will be copied.
-/// - `min_alignment` is the minimum alignment to which the copy will be aligned. The
-/// copy may not actually be aligned to `min_alignment` depending on the alignment requirements
-/// of `T` (the actual alignment will be the greater between `align_of::<T>` and `min_align.next_power_of_two()`).
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_to_offset_with_align_exact<T: Copy, S: Slab>(
-    src: &T,
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<CopyRecord, CopyError> {
-    let t_layout = Layout::new::<T>();
-    let record = compute_offsets(&*dst, start_offset, t_layout, min_alignment, true)?;
-
-    // SAFETY: if compute_offsets succeeded, this has already been checked to be safe.
-    let dst_ptr = unsafe { dst.base_ptr_mut().add(record.copy_start_offset) }.cast::<T>();
-
-    // SAFETY:
-    // - src is valid as we have a reference to it
-    // - dst is valid so long as requirements for `slab` were met, i.e.
-    // we have unique access to the region described and that it is valid for the duration
-    // of 'a.
-    // - areas not overlapping as long as safety requirements of creation of `self` were met,
-    // i.e. that we have exclusive access to the region of memory described.
-    // - dst aligned at least to align_of::<T>()
-    // - checked that copy stays within bounds of our allocation
-    unsafe {
-        core::ptr::copy_nonoverlapping(src as *const T, dst_ptr, 1);
-    }
-
-    Ok(record)
-}
-
-/// Copies `src` into the memory represented by `dst` starting at a minimum location
-/// of `start_offset` bytes past the start of `dst`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`,
-/// in bytes, before which any copied data will *certainly not* be placed. However,
-/// the actual beginning of the copied data may not be exactly at `start_offset` if
-/// padding bytes are needed to satisfy alignment requirements. The actual beginning
-/// of the copied bytes is contained in the returned [`CopyRecord`].
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_to_offset<T: Copy, S: Slab>(
-    src: &T,
-    dst: &mut S,
-    start_offset: usize,
-) -> Result<CopyRecord, CopyError> {
-    copy_to_offset_with_align(src, dst, start_offset, 1)
-}
-
-/// Copies `src` into the memory represented by `dst` starting at a minimum location
-/// of `start_offset` bytes past the start of `dst` and with minimum alignment
-/// `min_alignment`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`,
-/// in bytes, before which any copied data will *certainly not* be placed. However,
-/// the actual beginning of the copied data may not be exactly at `start_offset` if
-/// padding bytes are needed to satisfy alignment requirements. The actual beginning
-/// of the copied bytes is contained in the returned [`CopyRecord`].
-/// - `min_alignment` is the minimum alignment to which the copy will be aligned. The
-/// copy may not actually be aligned to `min_alignment` depending on the alignment requirements
-/// of `T` (the actual alignment will be the greater between `align_of::<T>` and `min_align.next_power_of_two()`).
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_to_offset_with_align<T: Copy, S: Slab>(
-    src: &T,
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<CopyRecord, CopyError> {
-    let t_layout = Layout::new::<T>();
-    let record = compute_offsets(&*dst, start_offset, t_layout, min_alignment, false)?;
-
-    // SAFETY: if compute_offsets succeeded, this has already been checked to be safe.
-    let dst_ptr = unsafe { dst.base_ptr_mut().add(record.copy_start_offset) }.cast::<T>();
-
-    // SAFETY:
-    // - src is valid as we have a reference to it
-    // - dst is valid so long as requirements for `slab` were met, i.e.
-    // we have unique access to the region described and that it is valid for the duration
-    // of 'a.
-    // - areas not overlapping as long as safety requirements of creation of `self` were met,
-    // i.e. that we have exclusive access to the region of memory described.
-    // - dst aligned at least to align_of::<T>()
-    // - checked that copy stays within bounds of our allocation
-    unsafe {
-        core::ptr::copy_nonoverlapping(src as *const T, dst_ptr, 1);
-    }
-
-    Ok(record)
-}
-
-/// Copies from `slice` into the memory represented by `dst` starting at *exactly*
-/// `start_offset` bytes past the start of `self`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`, in bytes,
-/// where the first byte of the copied data will be placed. If the requested
-/// start offset does not satisfy computed alignment requirements, an error will
-/// be returned and no data will be copied.
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_from_slice_to_offset_exact<T: Copy, S: Slab>(
-    src: &[T],
-    dst: &mut S,
-    start_offset: usize,
-) -> Result<CopyRecord, CopyError> {
-    copy_from_slice_to_offset_with_align(src, dst, start_offset, 1)
-}
-
-/// Copies from `slice` into the memory represented by `dst` starting at *exactly*
-/// `start_offset` bytes past the start of `dst` and with minimum alignment `min_alignment`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`, in bytes,
-/// where the first byte of the copied data will be placed. If the requested
-/// start offset does not satisfy computed alignment requirements, an error will
-/// be returned and no data will be copied.
-/// - `min_alignment` is the minimum alignment to which the copy will be aligned. The
-/// copy may not actually be aligned to `min_alignment` depending on the alignment requirements
-/// of `T` (the actual alignment will be the greater between `align_of::<T>` and `min_align.next_power_of_two()`).
-///     - The whole data of the slice will be copied directly, so, alignment between elements
-/// ignores `min_alignment`.
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_from_slice_to_offset_with_align_exact<T: Copy, S: Slab>(
-    src: &[T],
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<CopyRecord, CopyError> {
-    let t_layout = Layout::for_value(src);
-    let record = compute_offsets(&*dst, start_offset, t_layout, min_alignment, true)?;
-
-    // SAFETY: if compute_offsets succeeded, this has already been checked to be safe.
-    let dst_ptr = unsafe { dst.base_ptr_mut().add(record.copy_start_offset) }.cast::<T>();
-
-    // SAFETY:
-    // - src is valid as we have a reference to it
-    // - dst is valid so long as requirements for `slab` were met, i.e.
-    // we have unique access to the region described and that it is valid for the duration
-    // of 'a.
-    // - areas not overlapping as long as safety requirements of creation of `self` were met,
-    // i.e. that we have exclusive access to the region of memory described.
-    // - dst aligned at least to align_of::<T>()
-    // - checked that copy stays within bounds of our allocation
-    unsafe {
-        core::ptr::copy_nonoverlapping(src.as_ptr(), dst_ptr, src.len());
-    }
-
-    Ok(record)
-}
-
-/// Copies from `slice` into the memory represented by `dst` starting at a minimum location
-/// of `start_offset` bytes past the start of `self`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`,
-/// in bytes, before which any copied data will *certainly not* be placed. However,
-/// the actual beginning of the copied data may not be exactly at `start_offset` if
-/// padding bytes are needed to satisfy alignment requirements. The actual beginning
-/// of the copied bytes is contained in the returned [`CopyRecord`].
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_from_slice_to_offset<T: Copy, S: Slab>(
-    src: &[T],
-    dst: &mut S,
-    start_offset: usize,
-) -> Result<CopyRecord, CopyError> {
-    copy_from_slice_to_offset_with_align(src, dst, start_offset, 1)
-}
-
-/// Copies from `slice` into the memory represented by `dst` starting at a minimum location
-/// of `start_offset` bytes past the start of `dst`.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`,
-/// in bytes, before which any copied data will *certainly not* be placed. However,
-/// the actual beginning of the copied data may not be exactly at `start_offset` if
-/// padding bytes are needed to satisfy alignment requirements. The actual beginning
-/// of the copied bytes is contained in the returned [`CopyRecord`].
-/// - `min_alignment` is the minimum alignment to which the copy will be aligned. The
-/// copy may not actually be aligned to `min_alignment` depending on the alignment requirements
-/// of `T` (the actual alignment will be the greater between `align_of::<T>` and `min_align.next_power_of_two()`).
-///     - The whole data of the slice will be copied directly, so, alignment between elements
-/// ignores `min_alignment`.
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[inline]
-pub fn copy_from_slice_to_offset_with_align<T: Copy, S: Slab>(
-    src: &[T],
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<CopyRecord, CopyError> {
-    let t_layout = Layout::for_value(src);
-    let record = compute_offsets(&*dst, start_offset, t_layout, min_alignment, false)?;
-
-    // SAFETY: if compute_offsets succeeded, this has already been checked to be safe.
-    let dst_ptr = unsafe { dst.base_ptr_mut().add(record.copy_start_offset) }.cast::<T>();
-
-    // SAFETY:
-    // - src is valid as we have a reference to it
-    // - dst is valid so long as requirements for `slab` were met, i.e.
-    // we have unique access to the region described and that it is valid for the duration
-    // of 'a.
-    // - areas not overlapping as long as safety requirements of creation of `self` were met,
-    // i.e. that we have exclusive access to the region of memory described.
-    // - dst aligned at least to align_of::<T>()
-    // - checked that copy stays within bounds of our allocation
-    unsafe {
-        core::ptr::copy_nonoverlapping(src.as_ptr(), dst_ptr, src.len());
-    }
-
-    Ok(record)
-}
-
-/// Copies from `src` iterator into the memory represented by `dst` starting at a minimum location
-/// of `start_offset` bytes past the start of `dst`.
-///
-/// Returns a vector of [`CopyRecord`]s, one for each item in the `src` iterator.
-///
-/// - `start_offset` is the offset into the allocation represented by `dst`,
-/// in bytes, before which any copied data will *certainly not* be placed. However,
-/// the actual beginning of the copied data may not be exactly at `start_offset` if
-/// padding bytes are needed to satisfy alignment requirements. The actual beginning
-/// of the copied bytes is contained in the returned [`CopyRecord`]s.
-/// - `min_alignment` is the minimum alignment to which the copy will be aligned. The
-/// copy may not actually be aligned to `min_alignment` depending on the alignment requirements
-/// of `T` (the actual alignment will be the greater between `align_of::<T>` and `min_align.next_power_of_two()`).
-/// - For this variation, `min_alignment` will also be respected *between* elements yielded by
-/// the iterator. To copy inner elements aligned only to `align_of::<T>()` (i.e. with the layout of
-/// an `[T]`), see [`copy_from_iter_to_offset_with_align_packed`].
-///
-/// # Safety
-///
-/// This function is safe on its own, however it is very possible to do unsafe
-/// things if you read the copied data in the wrong way. See the
-/// [crate-level Safety documentation][`crate#safety`] for more.
-#[cfg(feature = "std")]
-#[inline]
-pub fn copy_from_iter_to_offset_with_align<T: Copy, Iter: Iterator<Item = T>, S: Slab>(
-    src: Iter,
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<Vec<CopyRecord>, CopyError> {
-    let mut offset = start_offset;
-
-    src.map(|item| {
-        let copy_record = copy_to_offset_with_align(&item, dst, offset, min_alignment)?;
-        offset = copy_record.copy_end_offset;
-        Ok(copy_record)
-    })
-    .collect::<Result<Vec<_>, _>>()
-}
-
-/// Like [`copy_from_iter_to_offset_with_align`] except that
-/// alignment between elements yielded by the iterator will ignore `min_alignment`
-/// and rather only be aligned to the alignment of `T`.
-///
-/// Because of this, only one [`CopyRecord`] is returned specifying the record of the
-/// entire block of copied data. If the `src` iterator is empty, returns `None`.
-#[inline]
-pub fn copy_from_iter_to_offset_with_align_packed<T: Copy, Iter: Iterator<Item = T>, S: Slab>(
-    mut src: Iter,
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<Option<CopyRecord>, CopyError> {
-    let first_record = if let Some(first_item) = src.next() {
-        copy_to_offset_with_align(&first_item, dst, start_offset, min_alignment)?
-    } else {
-        return Ok(None);
-    };
-
-    let mut prev_record = first_record;
-
-    for item in src {
-        let copy_record = copy_to_offset_with_align(&item, dst, prev_record.copy_end_offset, 1)?;
-        prev_record = copy_record;
-    }
-
-    Ok(Some(CopyRecord {
-        copy_start_offset: first_record.copy_start_offset,
-        copy_end_offset: prev_record.copy_end_offset,
-        copy_end_offset_padded: prev_record.copy_end_offset_padded,
-    }))
-}
-
-/// Like [`copy_from_iter_to_offset_with_align_packed`] except that it will return an error
-/// and no data will be copied if the supplied `start_offset` doesn't meet the computed alignment
-/// requirements.
-#[inline]
-pub fn copy_from_iter_to_offset_with_align_exact_packed<
-    T: Copy,
-    Iter: Iterator<Item = T>,
-    S: Slab,
->(
-    mut src: Iter,
-    dst: &mut S,
-    start_offset: usize,
-    min_alignment: usize,
-) -> Result<Option<CopyRecord>, CopyError> {
-    let first_record = if let Some(first_item) = src.next() {
-        copy_to_offset_with_align_exact(&first_item, dst, start_offset, min_alignment)?
-    } else {
-        return Ok(None);
-    };
-
-    let mut prev_record = first_record;
-
-    for item in src {
-        let copy_record =
-            copy_to_offset_with_align_exact(&item, dst, prev_record.copy_end_offset, 1)?;
-        prev_record = copy_record;
-    }
-
-    Ok(Some(CopyRecord {
-        copy_start_offset: first_record.copy_start_offset,
-        copy_end_offset: prev_record.copy_end_offset,
-        copy_end_offset_padded: prev_record.copy_end_offset_padded,
-    }))
 }